minor updates to welcome screen
