+++ app/src/main/java/com/hover/stax/onboarding/welcome/WelcomeFragment.kt
@@ -70,59 +70,6 @@ class WelcomeFragment : Fragment() {
         }
     }
 
-    companion object {
-        const val SALUTATIONS = "greetings"
-    }
-}
-
-
-//This is only useful when you're building and want to see what the design looks like. It can be removed in the final version
-@Preview
-@Composable
-fun WelcomeScreenPreview() {
-    StaxTheme {
-        Surface(
-            modifier = Modifier.fillMaxSize(),
-            color = MaterialTheme.colors.background
-        ) {
-            Scaffold(
-                modifier = Modifier.padding(24.dp),
-                content = { innerPadding ->
-                    Column(modifier = Modifier.padding(innerPadding)) {
-                        WelcomeHeader(
-                            title = stringResource(R.string.welcome_title_one),
-                            desc = stringResource(R.string.welcome_sub_one)
-                        )
-
-                        Spacer(modifier = Modifier.height(16.dp))
-
-                        Column {
-                            FeatureCard(
-                                title = stringResource(R.string.intro_feature_one_title),
-                                desc = stringResource(R.string.intro_feature_one_desc),
-                                iconRes = R.drawable.ic_automated
-                            )
-
-                            FeatureCard(
-                                title = stringResource(R.string.intro_feature_two_title),
-                                desc = stringResource(R.string.intro_feature_two_desc),
-                                iconRes = R.drawable.ic_control
-                            )
-
-                            FeatureCard(
-                                title = stringResource(R.string.intro_feature_three_title),
-                                desc = stringResource(R.string.intro_feature_three_desc),
-                                iconRes = R.drawable.ic_safe
-                            )
-                        }
-                    }
-                },
-                bottomBar = {
-                    ContinueButton(text = stringResource(id = R.string.explore_btn_text), onClick = { })
-                }
-            )
-        }
-    }
 }
